+++ src/lib/3.adapters/storage/scoped-externals.repository.ts
@@ -6,8 +6,9 @@ import type { Version } from "lib/1.domain";
 const createScopedExternalsRepository = (
     config: StorageConfig,
 ): ForScopedExternalsStorage => {
-    const STORAGE: StorageEntry<ScopedExternals> = config.storage("scoped-externals", {  });
-    if (config.clearCache) STORAGE.clear();
+    const STORAGE: StorageEntry<ScopedExternals> = config.storage(config.storageNamespace)("scoped-externals", {});
+    
+    if (config.clearStorage) STORAGE.clear();
 
     const _cache: ScopedExternals = STORAGE.get() ?? {};